+++ src/features/assignment/tests/AssignEdit.stories.tsx
@@ -1,11 +1,11 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import '@toast-ui/editor/dist/toastui-editor.css';
-import AssignEdit from "@features/assignment/AssignEdit.tsx";
+import AssignEditPage from "@pages/assignment/AssignEditPage.tsx";
 
 
-const meta: Meta<typeof AssignEdit> = {
-    title: 'features/Assignment/AssignEdit',
-    component: AssignEdit,
+const meta: Meta<typeof AssignEditPage> = {
+    title: 'features/Assignment/AssignEditPage',
+    component: AssignEditPage,
     parameters: {
         paddings: {
             default: '2rem',
@@ -14,6 +14,6 @@ const meta: Meta<typeof AssignEdit> = {
 };
 
 export default meta;
-type Story = StoryObj<typeof AssignEdit>;
+type Story = StoryObj<typeof AssignEditPage>;
 
 export const Default: Story = {};
